Multimodal adapter MPS support

+++ llama/utils.py
@@ -9,7 +9,7 @@ import torch
 
 def sample_top_p(probs, p):
     probs_sort, probs_idx = torch.sort(probs, dim=-1, descending=True)
-    probs_sum = torch.cumsum(probs_sort, dim=-1)
+    probs_sum = torch.cumsum(probs_sort.float(), dim=-1).half()
     mask = probs_sum - probs_sort > p
     probs_sort[mask] = 0.0
     probs_sort.div_(probs_sort.sum(dim=-1, keepdim=True))
